azure_app/requirements.txt
Flask==3.0.3
Flask-SocketIO==5.4.1
azure-cosmos==4.7.0
azure-identity==1.19.0
gunicorn==23.0.0
eventlet==0.37.0
python-dotenv==1.0.1tqdm
selenium
pandas
lxml
flask-cors>=4.0.0
openai>=1.3.0